+++ dubbo-samples-http/src/main/java/org/apache/dubbo/samples/http/api/DemoService.java
@@ -17,22 +17,10 @@
  *
  */
 
-package com.alibaba.dubbo.samples.http;
+package org.apache.dubbo.samples.http.api;
 
-import com.alibaba.dubbo.samples.http.api.DemoService;
-import org.springframework.context.support.ClassPathXmlApplicationContext;
+public interface DemoService {
 
-/**
- * HttpConsumer
- */
-public class HttpConsumer {
+    String sayHello(String name);
 
-    public static void main(String[] args) throws Exception {
-        ClassPathXmlApplicationContext context = new ClassPathXmlApplicationContext(new String[]{"spring/http-consumer.xml"});
-        context.start();
-        DemoService demoService = (DemoService) context.getBean("demoService");
-        String result = demoService.sayHello("world");
-        System.out.println(result);
-        System.in.read();
-    }
 }
