+++ dubbo-samples-http/src/main/resources/spring/http-consumer.xml
@@ -28,6 +28,6 @@
 
     <dubbo:registry address="zookeeper://127.0.0.1:2181"/>
 
-    <dubbo:reference id="demoService" interface="com.alibaba.dubbo.samples.http.api.DemoService"/>
+    <dubbo:reference id="demoService" interface="org.apache.dubbo.samples.http.api.DemoService"/>
 
 </beans>
